requirements.txt
Flask==3.0.0
Flask-Mail==0.9.1
python-dotenv==1.0.0
torch==2.2.0
transformers==4.38.0
requests==2.31.0
Werkzeug==3.0.1
supabase==2.3.0
python-jose==3.3.0